removing the filter folder
