+++ test/Garnet.test/AttributeExtractorTests.cs
@@ -4,7 +4,7 @@
 using System;
 using System.Text;
 using Allure.NUnit;
-using Garnet.server.Vector.Filter;
+using Garnet.server;
 using NUnit.Framework;
 using NUnit.Framework.Legacy;
