+++ test/Garnet.test/ExprRunnerTests.cs
@@ -3,7 +3,7 @@
 
 using System.Text;
 using Allure.NUnit;
-using Garnet.server.Vector.Filter;
+using Garnet.server;
 using NUnit.Framework;
 using NUnit.Framework.Legacy;
